services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pymongo==4.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
Pillow==10.2.0
tensorflow==2.15.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
pytest==7.4.4
httpx==0.26.0
